Open Source library for Lynx motion arm AL5D for academic use on Linux
Initial commit.

Run `make` to compile demo.

If you want to create changes on the library, you can compile it with with
`make libufrn`. Just be careful with the absolute values that limits the
torque on the servo motors. Initial documentation is available in portu-
guese at the /doc/ folder.

Signed-off-by: Desnes A. Nunes do Rosario <desnesn@gmail.com>
Signed-off-by: Danilo Ichihara <daniloichihara@gmail.com>